Revert "Cache media (#6902)"
This reverts commit ca42f9b0ebfa1f4e8e86745a79af138b5865daee.

This is causing Chrome quota exceeded issues which is very hard to recover from.
