+++ app/javascript/mastodon/actions/importer/index.js
@@ -1,4 +1,4 @@
-import { putAccounts, putStatuses } from '../../storage/modifier';
+import { putAccounts, putStatuses } from '../../db/modifier';
 import { normalizeAccount, normalizeStatus } from './normalizer';
 
 export const ACCOUNT_IMPORT = 'ACCOUNT_IMPORT';
