+++ app/javascript/mastodon/service_worker/entry.js
@@ -1,10 +1,6 @@
 import './web_push_notifications';
 
-function openSystemCache() {
-  return caches.open('mastodon-system');
-}
-
-function openWebCache() {
+function openCache() {
   return caches.open('mastodon-web');
 }
 
@@ -15,7 +11,7 @@ function fetchRoot() {
 // Cause a new version of a registered Service Worker to replace an existing one
 // that is already installed, and replace the currently active worker on open pages.
 self.addEventListener('install', function(event) {
-  event.waitUntil(Promise.all([openWebCache(), fetchRoot()]).then(([cache, root]) => cache.put('/', root)));
+  event.waitUntil(Promise.all([openCache(), fetchRoot()]).then(([cache, root]) => cache.put('/', root)));
 });
 self.addEventListener('activate', function(event) {
   event.waitUntil(self.clients.claim());
@@ -25,7 +21,7 @@ self.addEventListener('fetch', function(event) {
 
   if (url.pathname.startsWith('/web/')) {
     const asyncResponse = fetchRoot();
-    const asyncCache = openWebCache();
+    const asyncCache = openCache();
 
     event.respondWith(asyncResponse.then(async response => {
       if (response.ok) {
@@ -35,10 +31,10 @@ self.addEventListener('fetch', function(event) {
       }
 
       throw null;
-    }).catch(() => asyncCache.then(cache => cache.match('/'))));
+    }).catch(() => caches.match('/')));
   } else if (url.pathname === '/auth/sign_out') {
     const asyncResponse = fetch(event.request);
-    const asyncCache = openWebCache();
+    const asyncCache = openCache();
 
     event.respondWith(asyncResponse.then(async response => {
       if (response.ok || response.type === 'opaqueredirect') {
@@ -47,22 +43,6 @@ self.addEventListener('fetch', function(event) {
       }
 
       return response;
-    }));
-  } else if (process.env.CDN_HOST ? url.host === process.env.CDN_HOST : url.pathname.startsWith('/system/')) {
-    event.respondWith(openSystemCache().then(async cache => {
-      const cached = await cache.match(event.request.url);
-
-      if (cached === undefined) {
-        const fetched = await fetch(event.request);
-
-        if (fetched.ok) {
-          await cache.put(event.request.url, fetched.clone());
-        }
-
-        return fetched;
-      }
-
-      return cached;
     }));
   }
 });
