+++ package.json
@@ -57,7 +57,6 @@
     "hammerjs": "^2.0.8",
     "http-link-header": "^0.8.0",
     "immutable": "^3.8.2",
-    "imports-loader": "^0.8.0",
     "intersection-observer": "^0.5.0",
     "intl": "^1.2.5",
     "intl-messageformat": "^2.2.0",
